+++ Runtime/ResourcesMusicLoader.cs
@@ -8,6 +8,11 @@ namespace Mane.SoundManeger
 {
     public class ResourcesMusicLoader : IMusicLoader
     {
+        /// <summary>
+        /// Resources loader doesn't need to retry as failures are permanent
+        /// </summary>
+        public bool ShouldRetry => false;
+        
         public async Task<AudioClip> GetMusicAsync(MonoBehaviour owner, string path, CancellationToken token = default)
         {
             if (string.IsNullOrEmpty(path)) return null;